Create destinations.py
This class is used to simulate the arrival of new vehicles to the traffic system. Each call to its step method returns either None, indicating that no vehicle came this step, or a destination ('W' or 'S').

The class could be based on measured or estimated traffic densities at different times of the day, but in our case it returns a predetermined sequence. This makes it easier to predict the result and thus also check that the program is correct.