Create traffic_components.py
Vehicle:
The Vehicle class is very simple. Its task is only to keep track of when the vehicle was created and where it is going. For this, two instance variables are sufficient: named destination and time of creation.

A vehicle does not need to know where it is, or to check traffic lights, or see other vehicles. The vehicles can be seen as small pawns moved by "someone else".

Thus, only the constructor is needed, that receives the destination and time of creation. If you want, you can add a __str__() method.
Lane:

A lane is a structure that contains a number of vehicles and a number of empty spaces. The vehicles can only enter at the beginning and leave at the end of the lane i.e. the lane is like a "tube":
Light:
This class is used to represent traffic signals.