Create traffic_system1.py
The traffic system will consist of three lanes (lane, lane_west, and lane_south) and two traffic lights: light_west and light_south (previously we called them s1 and s2).
It can be nice to have short lanes after (to the left of) the traffic lights so that you can see how the vehicles pass.

At point E there is a queue (not drawn).

At a time step, one or more of the following things may happen, in this order:

A vehicle from lane_west and/or lane_south passes its signal (if it is green).
The lane_west and lane_south files are stepped.
A vehicle immediately to the right of X is moved to lane_west or lane_south depending on its destination (W or S). If the destination lane is not free, the vehicle remains where it is and the split point is counted as blocked.
The vehicles on lane are stepped.
A vehicle arrives to the system at point E and is queued.
If the last place in lane is vacant and there are vehicles in the queue, the first from the queue is taken and put to the last place in lane.
The light signals are stepped.
A simulation thus consists of a time stepping scheme where the above things happen. It is appropriate that the steps, just as in reality, are performed in the order in which they appear above. It can be practical to let your step method in turn call different methods that you introduce yourself to take care of different parts. It can reduce the amount of repetitive code in steps and make it clearer.